task_manager/requirements.txt
Flask==2.0.1
Flask-Cors==4.0.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
typing-extensions>=4.7.1
openai==0.28.1
tiktoken==0.5.1
azure-identity==1.15.0
azure-core==1.29.5
python-dateutil==2.8.2
requests==2.31.0
# Nuevas dependencias para MySQL y Pydantic
pymysql==1.1.0
mysql-connector-python==8.2.0
pydantic==2.5.0
alembic==1.13.1
cryptography>=42.0.0

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0
pytest-html==4.1.1
coverage==7.3.2